# Switch Reed-Solomon implementation to github.com/klauspost/reedsolomon