--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -26,20 +26,12 @@ jobs:
             **/**.go
             go.mod
             go.sum
-      - name: install leopard
-        run: |
-          git clone https://github.com/celestiaorg/go-leopard.git
-          pushd go-leopard
-          make build-cleo
-          sudo make install-cleo
-          popd
-        if: env.GIT_DIFF
       - name: build
-        run: GOOS=linux GOARCH=${{ matrix.goarch }} go build -tags leopard
+        run: GOOS=linux GOARCH=${{ matrix.goarch }} go build
         if: env.GIT_DIFF
       - name: test & coverage report creation
         run: |
-          GOARCH=${{ matrix.goarch }} go test -tags leopard -mod=readonly -timeout 8m -race -coverprofile=coverage.txt -covermode=atomic
+          GOARCH=${{ matrix.goarch }} go test -mod=readonly -timeout 8m -race -coverprofile=coverage.txt -covermode=atomic
         if: env.GIT_DIFF
       - uses: codecov/codecov-action@v1.0.15
         with: