--- a/go.mod
+++ b/go.mod
@@ -3,15 +3,16 @@ module github.com/celestiaorg/rsmt2d
 go 1.17
 
 require (
-	// only needed if built with go build -tags leopard
-	github.com/celestiaorg/go-leopard v0.1.0
 	github.com/celestiaorg/merkletree v0.0.0-20210714075610-a84dc3ddbbe4
 	github.com/stretchr/testify v1.7.0
 	github.com/vivint/infectious v0.0.0-20200605153912-25a574ae18a3
 )
 
+require github.com/klauspost/reedsolomon v1.10.0
+
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/klauspost/cpuid/v2 v2.0.14 // indirect
 	github.com/kr/text v0.2.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	gitlab.com/NebulousLabs/errors v0.0.0-20200929122200-06c536cf6975 // indirect