--- a/leopard.go
+++ b/leopard.go
@@ -1,11 +1,8 @@
-//go:build leopard
-
-// Note that if the build tag leopard is used, liblibleopard.a
-// has to be present where the linker will find it.
-// Otherwise go-leopard won't build.
 package rsmt2d
 
-import "github.com/celestiaorg/go-leopard"
+import (
+	"github.com/klauspost/reedsolomon"
+)
 
 var _ Codec = leoRSFF8Codec{}
 var _ Codec = leoRSFF16Codec{}
@@ -18,12 +15,37 @@ func init() {
 type leoRSFF8Codec struct{}
 
 func (l leoRSFF8Codec) Encode(data [][]byte) ([][]byte, error) {
-	return leopard.Encode(data)
+	return encode(data)
+}
+
+func encode(data [][]byte) ([][]byte, error) {
+	enc, err := reedsolomon.New(len(data), len(data))
+	if err != nil {
+		return nil, err
+	}
+	shards := make([][]byte, len(data)*2)
+	copy(shards, data)
+	for i := len(data); i < len(shards); i++ {
+		shards[i] = make([]byte, len(data[0]))
+	}
+	if err := enc.Encode(shards); err != nil {
+		return nil, err
+	}
+	return shards[len(data):], nil
 }
 
 func (l leoRSFF8Codec) Decode(data [][]byte) ([][]byte, error) {
+	return decode(data)
+}
+
+func decode(data [][]byte) ([][]byte, error) {
 	half := len(data) / 2
-	return leopard.Decode(data[:half], data[half:])
+	enc, err := reedsolomon.New(half, half)
+	if err != nil {
+		return nil, err
+	}
+	err = enc.Reconstruct(data)
+	return data, err
 }
 
 func (l leoRSFF8Codec) maxChunks() int {
@@ -37,12 +59,11 @@ func newLeoRSFF8Codec() leoRSFF8Codec {
 type leoRSFF16Codec struct{}
 
 func (leo leoRSFF16Codec) Encode(data [][]byte) ([][]byte, error) {
-	return leopard.Encode(data)
+	return encode(data)
 }
 
 func (leo leoRSFF16Codec) Decode(data [][]byte) ([][]byte, error) {
-	half := len(data) / 2
-	return leopard.Decode(data[:half], data[half:])
+	return decode(data)
 }
 
 func (leo leoRSFF16Codec) maxChunks() int {