--- a/rsmt2d_test.go
+++ b/rsmt2d_test.go
@@ -1,5 +1,3 @@
-//go:build leopard
-
 package rsmt2d_test
 
 import (