This is a work in progress until github/klauspost/reedsolomon is fast and stable enough to replace go-leopard. See https://github.com/klauspost/reedsolomon/pull/191 for the leopard port.

@liamsi what's the next step? Obviously performance optimizations to bring the Go port closer to the C++ version, but how much? I see references to the 8-bit version of leopard, do you need that ported as well or is klauspost/reedsolomon enough for total share counts <= 256?

CC @odeke-em .