@liamsi 2022-07-06
>  I see references to the 8-bit version of leopard, do you need that ported as well [...]

No, the 16-bit version of leopard should be sufficient IMO. 

Will review this PR and https://github.com/klauspost/reedsolomon/pull/191 shortly. Thanks a lot! Really excited about this!