@odeke-em 2022-07-21
@elias-orijtech thanks for working on this, nice work! @klauspost merged your changes per https://github.com/klauspost/reedsolomon/pull/191 so all good to go!