@musalbas 2022-09-09
According to these benchmarks by @evan-forbes, this new implementation is 40x slower than go-leopard in rsmt2d. Could there be an issue with the way it's integrated in rsmt2d?

```
BenchmarkRepair/Repairing_128x128_ODS_using_RSGF8-24              	       1	16736594416 ns/op
BenchmarkExtension/RSGF8_size_128-24                               	      16	  66631976 ns/op

new pure go leopard
BenchmarkRepair/Repairing_128x128_ODS_using_LeopardFF16-24         	       1	25971504223 ns/op
BenchmarkExtension/LeopardFF16_size_128-24                         	       1	3378071341 ns/op

old c++ leopard
BenchmarkRepair/Repairing_128x128_ODS_using_LeopardFF16-24        	       1	16808911720 ns/op
BenchmarkExtension/LeopardFF16_size_128-24                        	      40	  28827064 ns/op```