@musalbas 2022-09-12
For encoding, according to the profiler, most of the time for encoding is spent on calling New(). Maybe the same re-encoder can be reused instead of creating a new one every time? It says here it can be reused: https://github.com/klauspost/reedsolomon/blob/master/reedsolomon.go#L387

![profile001](https://user-images.githubusercontent.com/5038597/189629376-49dca0bd-ce8f-48af-aa9c-947054395703.gif)