@klauspost 2022-09-12
Yes, encoders can/should be reused. Forward matrices are preserved between calls, and for gf8 inverse matrices as well. The encoder can safely be shared between goroutines.

You can [force GF16](https://github.com/klauspost/reedsolomon#forcing-leopard-gf16) for all sizes, but reconstruction is rather expensive.